# Refactor : 캐러셀 폴더 생성 및 정리
